pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "faiss-cpu>=1.10.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.4",
    "gunicorn>=23.0.0",
    "langchain-community>=0.3.18",
    "langchain>=0.3.19",
    "langchain-google-genai>=2.0.10",
    "psycopg2-binary>=2.9.10",
    "pypdf>=5.3.0",
    "requests>=2.32.3",
]
